--- src/app/(home)/movieList.tsx
@@ -1,9 +1,0 @@
-import axios from "axios";
-
-async function getMovie() {
-  return await axios.get("movie/popular");
-}
-
-export default function MovieList() {
-  return <div>movieList</div>;
-}
