+++ src/app/(home)/page.tsx
@@ -1,5 +1,15 @@
-import MovieList from "./movieList";
+import { Metadata } from "next";
+import MovieList from "./(movie)/movieList";
+
+export const metadata: Metadata = {
+  title: "나의 영화 일지",
+  description: "영화 기록하기",
+};
 
 export default function Home() {
-  return <MovieList />;
+  return (
+    <div>
+      <MovieList></MovieList>
+    </div>
+  );
 }
