+++ src/app/layout.tsx
@@ -2,6 +2,11 @@ import type { Metadata } from "next";
 import "../styles/globals.css";
 import ScrollBtn from "./(common)/scrollBtn";
 import Nav from "./(home)/nav";
+import ReduxProvider from "../provider/redux";
+import MuiProvider from "../provider/mui";
+import QueryProvider from "@/provider/query";
+import SessionProvider from "@/provider/session";
+import { Session } from "next-auth";
 
 export const metadata: Metadata = {
   title: "나의 영화 일지",
@@ -10,15 +15,25 @@ export const metadata: Metadata = {
 
 export default function RootLayout({
   children,
+  session,
 }: Readonly<{
   children: React.ReactNode;
+  session: Session;
 }>) {
   return (
     <html lang="kr">
       <body>
-        <Nav />
-        {children}
-        <ScrollBtn />
+        <ReduxProvider>
+          <QueryProvider>
+            <SessionProvider session={session}>
+              <MuiProvider>
+                <Nav />
+                {children}
+                <ScrollBtn />
+              </MuiProvider>
+            </SessionProvider>
+          </QueryProvider>
+        </ReduxProvider>
       </body>
     </html>
   );
