--- src/pages/_document.tsx
@@ -1,36 +1,0 @@
-import {
-  Head,
-  Main,
-  NextScript,
-  DocumentProps,
-  DocumentContext,
-  Html,
-} from "next/document";
-import {
-  DocumentHeadTags,
-  DocumentHeadTagsProps,
-  documentGetInitialProps,
-} from "@mui/material-nextjs/v13-pagesRouter";
-
-export default function Document(props: DocumentProps & DocumentHeadTagsProps) {
-  return (
-    <Html>
-      <Head>
-        <DocumentHeadTags {...props} />
-        <link
-          rel="stylesheet"
-          href="https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@100..900&display=swap"
-        />
-      </Head>
-      <body>
-        <Main />
-        <NextScript />
-      </body>
-    </Html>
-  );
-}
-
-Document.getInitialProps = async (ctx: DocumentContext) => {
-  const finalProps = await documentGetInitialProps(ctx);
-  return finalProps;
-};
