--- src/pages/calendar/index.tsx
@@ -1,14 +1,0 @@
-import React, { useState } from "react";
-
-import CalendarMap from "@/components/calendar/CalendarMap";
-
-/**
- * @returns calendar page
- */
-export default function Index() {
-  return (
-    <React.Fragment>
-      <CalendarMap></CalendarMap>
-    </React.Fragment>
-  );
-}
