+++ src/styles/globals.css
@@ -17,7 +17,8 @@
 }
 
 body {
-  background-color: #0f1015;
+  color: "black";
+  /* background-color: #0f1015; */
 }
 
 html,
